Modified planet visibility and added enemy image
I made it so that planets are invisible by default and are set to be
visible when they are first generated. This is to prevent them from
being rendered before all of their fields are set properly, an issue
which has not occurred yet but may in future versions. Enemies will be
implemented shortly.