Added layering and enemies and updated background
I added in layering to control which objects are rendered on top of the
others in order to ensure enemies don't disappear under planets. Also, I
added enemies, with an AI that allows them to function normally except
they cannot shoot. Also, they cannot be shot yet. And I put in the
background that Shelby made.