Rearranged some stuff to make stuff run better
The main thing I did was put an update method in all entities to be used
in place of their update code being in the region class. I also cleaned
up the code for rendering the cats so that it didn't add and
subsequently subtract the screen position, which made it significantly
less buggy. I attempted to implement a better system of isolating
entities of a given class by storing them in a map with their class as
the key, but given the frequent need to cycle through all entities, it
proved impractical and harmed performance, so I removed it, but there
may be artifacts. Also I replaced the use of CopyOnWrightArrayList for
thread-safing of the entities list with a thread-safe arraylist. I don't
know if it helped performance, but it certainly didn't hurt it.